+++ packages/ui/src/molecules/ItemCard/index.ts
@@ -1,0 +1,2 @@
+export { ItemCard } from './ItemCard'
+export type { ItemCardProps, ItemCardVariant, ItemCardStatus } from './ItemCard.types'