+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vacacode/ui
 
+## 0.1.0
+
+### Minor Changes
+
+- Add ItemCard molecule and its related elements:
+  - ItemCard story.
+  - ItemCard test.
+  - Atoms:
+    - Badge.
+    - Skeleton.
+    - Thumbnail.
+
 ## 0.0.2
 
 ### Patch Changes